<state_machine> Fix warning

+++ state/lib/state_machine.c
@@ -95,7 +95,7 @@ static inline int state_machine_call_event_normal(StateMachine this, int event, 
 /*! for multi thread, call event */
 static int state_machine_call_event_multithread(StateMachine this, int event, void *arg, int arglen, void (*response)(int result));
 /*! for multi thread, thread main callback for threadpool */
-static void state_machine_thread_main(int socketfd, short eventflag, void * event_arg);
+static void state_machine_thread_main(int socketfd, int eventflag, void * event_arg);
 /*! for multi thread, open socket */
 static inline int state_machine_open_socket(StateMachine this);
 /*! for multi thread, close socket */
@@ -245,7 +245,7 @@ static int state_machine_call_event_multithread(StateMachine this, int event, vo
 }
 
 /*! for multi thread, thread main */
-static void state_machine_thread_main(int socketfd, short eventflag, void * event_arg) {
+static void state_machine_thread_main(int socketfd, int eventflag, void * event_arg) {
 	StateMachine this = (StateMachine)event_arg;
 	int ret = 0;
 	StateMachineMsg msg;
